PP-5443 Moves validaiton tests from GetTelephonePaymentIT into BuiltInValidationIT
