+++ src/test/java/uk/gov/pay/api/it/telephone/BuiltInValidationIT.java
@@ -1,7 +1,9 @@
 package uk.gov.pay.api.it.telephone;
 
+import org.apache.commons.lang.StringUtils;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Test;
 import uk.gov.pay.api.utils.PublicAuthMockClient;
 
 import java.util.Map;
@@ -30,5 +32,46 @@ public class BuiltInValidationIT extends TelephonePaymentResourceITBase {
     public void tearDown() {
         requestBody.clear();
     }
-    
+
+    @Test
+    public void respondWith422_whenReferenceLengthIsGreaterThanMaxValue() {
+        requestBody.replace("reference", StringUtils.repeat("*", 256));
+        postPaymentResponse(toJson(requestBody))
+                .statusCode(422);
+    }
+
+    @Test
+    public void respondWith422_whenDescriptionLengthIsGreaterThanMaxValue() {
+        requestBody.replace("description", StringUtils.repeat("*", 256));
+        postPaymentResponse(toJson(requestBody))
+                .statusCode(422);
+    }
+
+    @Test
+    public void respondWith422_whenProcessorIdIsMissing() {
+        requestBody.remove("processor_id");
+        postPaymentResponse(toJson(requestBody))
+                .statusCode(422);
+    }
+
+    @Test
+    public void respondWith422_whenProcessorIdIsNull() {
+        requestBody.replace("processor_id", null);
+        postPaymentResponse(toJson(requestBody))
+                .statusCode(422);
+    }
+
+    @Test
+    public void respondWith422_whenProviderIdIsMissing() {
+        requestBody.remove("provider_id");
+        postPaymentResponse(toJson(requestBody))
+                .statusCode(422);
+    }
+
+    @Test
+    public void respondWith422_whenProviderIdIsNull() {
+        requestBody.replace("provider_id", null);
+        postPaymentResponse(toJson(requestBody))
+                .statusCode(422);
+    }
 }
